+++ R/find_formula.R
@@ -180,6 +180,12 @@ find_formula.gamm <- function(x, ...) {
 
 
 #' @export
+find_formula.rma <- function(x, ...) {
+  NULL
+}
+
+
+#' @export
 find_formula.gee <- function(x, ...) {
   tryCatch(
     {
@@ -481,6 +487,27 @@ find_formula.feis <- function(x, ...) {
 
 
 #' @export
+find_formula.bife <- function(x, ...) {
+  f <- .safe_deparse(stats::formula(x))
+  f_parts <- unlist(strsplit(f, "|", fixed = TRUE))
+
+  f.cond <- .trim(f_parts[1])
+
+  if (length(f_parts) > 1) {
+    f.rand <- paste0("~", .trim(f_parts[2]))
+  } else {
+    f.rand <- NULL
+  }
+
+  .compact_list(list(
+    conditional = stats::as.formula(f.cond),
+    random = stats::as.formula(f.rand)
+  ))
+}
+
+
+
+#' @export
 find_formula.wbm <- function(x, ...) {
   f <- .safe_deparse(stats::formula(x))
   f_parts <- unlist(strsplit(f, "(?<!\\()\\|(?![\\w\\s\\+\\(~]*[\\)])", perl = TRUE))
@@ -572,6 +599,13 @@ find_formula.zeroinfl <- find_formula.hurdle
 
 #' @export
 find_formula.zerotrunc <- find_formula.hurdle
+
+
+#' @export
+find_formula.zcpglm <- function(x, ...) {
+  .zeroinf_formula(x, separator = "\\|\\|")
+}
+
 
 
 
@@ -1038,7 +1072,7 @@ find_formula.BFBayesFactor <- function(x, ...) {
 
 # Find formula for zero-inflated regressions, where
 # zero-inflated part is separated by | from count part
-.zeroinf_formula <- function(x) {
+.zeroinf_formula <- function(x, separator = "\\|") {
   f <- tryCatch(
     {
       stats::formula(x)
@@ -1052,7 +1086,7 @@ find_formula.BFBayesFactor <- function(x, ...) {
     return(NULL)
   }
 
-  f <- .trim(unlist(strsplit(.safe_deparse(f), "\\|")))
+  f <- .trim(unlist(strsplit(.safe_deparse(f), separator)))
 
   c.form <- stats::as.formula(f[1])
   if (length(f) == 2) {