+++ R/get_statistic.R
@@ -82,6 +82,11 @@ get_statistic.lme <- function(x, ...) {
   get_statistic.default(x, column_index = 4)
 }
 
+#' @export
+get_statistic.lmerModLmerTest <- get_statistic.lme
+
+
+
 
 #' @export
 get_statistic.plm <- get_statistic.default
@@ -167,9 +172,17 @@ get_statistic.zeroinfl <- function(x, component = c("all", "conditional", "zi", 
   cs <- .compact_list(stats::coef(summary(x)))
   out <- lapply(names(cs), function(i) {
     comp <- ifelse(i == "count", "conditional", "zi")
+    stats <- cs[[i]]
+
+    # remove log(theta)
+    theta <- grepl("Log(theta)", rownames(stats), fixed = TRUE)
+    if (any(theta)) {
+      stats <- stats[!theta, ]
+    }
+
     data.frame(
       Parameter = find_parameters(x, effects = "fixed", component = comp, flatten = TRUE),
-      Statistic = as.vector(cs[[i]][, 3]),
+      Statistic = as.vector(stats[, 3]),
       Component = comp,
       stringsAsFactors = FALSE,
       row.names = NULL
@@ -658,6 +671,42 @@ get_statistic.cpglm <- function(x, ...) {
 
 
 
+#' @importFrom utils capture.output
+#' @export
+get_statistic.zcpglm <- function(x, component = c("all", "conditional", "zi", "zero_inflated"), ...) {
+  if (!requireNamespace("cplm", quietly = TRUE)) {
+    stop("To use this function, please install package 'cplm'.")
+  }
+
+  component <- match.arg(component)
+  junk <- utils::capture.output(stats <- cplm::summary(x)$coefficients)
+  params <- get_parameters(x)
+
+  tweedie <- data.frame(
+    Parameter = params$Parameter[params$Component == "conditional"],
+    Statistic = as.vector(stats$tweedie[, "z value"]),
+    Component = "conditional",
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+
+  zero <- data.frame(
+    Parameter = params$Parameter[params$Component == "zero_inflated"],
+    Statistic = as.vector(stats$zero[, "z value"]),
+    Component = "zero_inflated",
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+
+  out <- .filter_component(rbind(tweedie, zero), component)
+  out <- .remove_backticks_from_parameter_names(out)
+
+  attr(out, "statistic") <- find_statistic(x)
+  out
+}
+
+
+
 #' @export
 get_statistic.MANOVA <- function(x, ...) {
   stats <- as.data.frame(x$WTS)
@@ -1069,3 +1118,23 @@ get_statistic.rma <- function(x, ...) {
   attr(out, "statistic") <- find_statistic(x)
   out
 }
+
+
+
+#' @export
+get_statistic.bife <- function(x, ...) {
+  parms <- get_parameters(x)
+  cs <- summary(x)
+
+  out <- data.frame(
+    Parameter = parms$Parameter,
+    Statistic = as.vector(cs$cm[, 3]),
+    stringsAsFactors = FALSE,
+    row.names = NULL
+  )
+
+  attr(out, "statistic") <- find_statistic(x)
+  out
+}
+
+