+++ R/n_obs.R
@@ -189,6 +189,11 @@ n_obs.cpglm <- function(x, ...) {
   nrow(x$model.frame)
 }
 
+#' @export
+n_obs.zcpglm <- n_obs.cpglm
+
+#' @export
+n_obs.bcplm <- n_obs.cpglm
 
 
 #' @export
@@ -290,9 +295,11 @@ n_obs.fixest <- function(x, ...) {
 
 #' @export
 n_obs.feglm <- function(x, ...) {
-  x$nobs["nobs"]
+  x$nobs[["nobs"]]
 }
 
+#' @export
+n_obs.bife <- n_obs.feglm
 
 #' @export
 n_obs.complmrob <- n_obs.cgam